+++ test/dummy/person.rb
@@ -1,5 +1,6 @@
 class Person
   include Mongoid::Document
+  include Mongoid::Paranoia
 
   field :name, type: String, default: "me"
 end