+++ test/matchers/document_test.rb
@@ -4,4 +4,5 @@ describe "Document matchers" do
   subject { Person }
 
   it { must be_document }
+  it { must be_paranoid }
 end