+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ### 0.0.1
 
+* Added `be_paranoid` document matcher.
 * Added default failure messages based on description matcher.
-* Added `be_document` matcher.
+* Added `be_document` document matcher.
 * Added `with_default_value` method to `HaveFieldMatcher`.
 * Added `of_type` method to `HaveFieldMatcher`.
 * Added `must` and `wont` methods to `MiniTest::Spec`.